reuse max borrow logic for manage pos and create pos
